requirements.txt
# Core ML stack
transformers==4.44.2
accelerate==0.34.2
peft==0.12.0
unsloth
trl==0.9.6

# MLX for Apple Silicon
mlx==0.29.1
mlx-lm==0.28.0

# Data and validation
pydantic==2.9.2
pandas==2.2.2
pyarrow==16.1.0
orjson==3.10.7

# Tracking and monitoring
wandb==0.17.9
weave==0.50.8
psutil==6.0.0
tensorboard==2.16.2

# Data processing (no Snowflake dependencies)

# Evaluation
lm-eval==0.4.2

# CLI / Utils
typer==0.12.5
rich==13.8.0
python-dotenv==1.0.1

# Testing
pytest==8.3.3
pytest-cov==5.0.0
